fix: filter couches without fetching with JS (#181)
* fix: filter couches without fetching with JS

* validate development users to have complete profiles

* show user pictures in local

* remove timeout

* add hidden input for show/hide filters
