--- app/views/couches/index.json.jbuilder
@@ -1,1 +1,0 @@
-json.inserted_list render(partial: 'couches/list', formats: :html, locals: { couches: @couches })
