+++ app/views/partials/_couches.html.erb
@@ -1,5 +1,5 @@
-<div data-characteristics-filter-target="couches">
-  <ul class="couches__list" data-characteristics-filter-target="couches">
+<div>
+  <ul class="couches__list">
     <% couches.each do |couch| %>
       <%= render 'partials/couch', couch: couch %>
     <% end %>
